Fix `save_json()` serialization issue
This commit substantively fixes the `save_json()` serialization issue
and removes the workaround introduced by the previous commit. When using
numeric labels, `save_json()` fails because the labels are a
`torch.Tensor` dtype which is not serializable by the `json` library.
This commit adds some logic to use `torch.Tensor.item()` to return the
label as a standard Python dtype which can then be handled by `json`.