Create `unittest`-ing harness in `tests/`
We create the foundation of a unit testing harness using the Python
module `unittest`, as well as update the project documentation with
instructions on how to run `mypy` for static type analysis and
`unittest` for unit testing. The tests are located in the `tests/`
directory, and mirror each file in `pytorch_influence_functions`. Be
advised that right now most of the tests are placeholders and are not
actually implemented.