Add Github Actions (GA) for CI/CD pipeline
Github Actions (GA) is a Continuous Integration/Continuous Delivery
pipeline that allows projects to automatically run tests and build
packages for release and distribution. This commit adds a GA workflow to
automatically run static analysis and unit tests, as well as GA-related
documentation.

This workflow does not currently build release packages or upload to
pip, but we hope to implement these as future steps once the package is
more mature.