Create GA workflow for builds and releases
This commit changes the `run-tests.yaml` to have hatch build the
package, and then creates a github release and uploads the build
artefacts. The GA workflow has to gather some environmental data such as
the package version, and the filenames and paths of the wheel and source
distributions in order to do so.